requirements.txt
fire==0.4.0
mysql-connector-python==8.0.25
protobuf==3.17.1
six==1.16.0
termcolor==1.1.0
pandas==1.2.4
numpy==1.20.3
PyYAML~=5.4.1